@@ -8,6 +8,7 @@ use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\AccueilController;
 use App\Http\Controllers\directorController;
 use App\Http\Controllers\NewsController;
+use App\Http\Controllers\teacherController;
 use App\Http\Controllers\UtilisationDuTempsController;
 use App\Http\Controllers\ContactController;
 use App\Http\Controllers\Contact2Controller;
@@ -41,15 +42,16 @@ Route::get('/logout', [LoginController::class, 'logout'])->name('logout');
 
 // });
 
+
+
+// director dashboard :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+
 Route::group(['middleware' => ['auth', DirectorMiddleware::class]], function () {
     // Route::get('/director/dashboard', [directorController::class, "directorDashboard"])->name('directorDashboard');
-
     // Route::get('/dashboard', [directorController::class, "generalGuardDashboard"])->name('generalGuard_dashboard');
 
     Route::get('/dashboard/allGeneralGuard', [directorController::class, "index"])->name('general_guard');
-    Route::get('/director/dashboard/addGeneralGuard', [directorController::class, "create"])->name('addGeneralGuard');
-
-
+    Route::get('/dashboard/addGeneralGuard', [directorController::class, "create"])->name('addGeneralGuard');
     Route::delete('/dashboard/deleteGeneralGuard/{id}', [directorController::class, "destroy"])->name('deleteGeneral_guard');
     Route::get('/dashboard/updateGeneralGuard/{id}/edit', [directorController::class, "edit"])->name('updateGeneral_guard');
     Route::patch('/dashboard/updateGeneralGuard/{id}', [directorController::class, "update"])->name('saveUpdate');
@@ -71,9 +73,12 @@ Route::group(['middleware' => ['auth', DirectorMiddleware::class]], function () 
 
 
 
-
-
-
+    Route::get('/dashboard/professeurs', [directorController::class, "displayTeachers"])->name('displayTeachers');
+    Route::get('/dashboard/addTeacher', [directorController::class, "addTeacher"])->name('addTeacher');
+    Route::get('/dashboard/updateTeacher/{id}', [directorController::class, "updateTeacher"])->name('updateTeacher');
+    Route::patch('/dashboard/updateTeacher/{id}', [directorController::class, "saveUpdate"])->name('saveUpdateTeacher');
+    Route::post('/dashboard/addTeacher', [directorController::class, "saveTeacher"])->name('saveTeacher');
+    Route::delete('/dashboard/deleteTeacher/{id}', [directorController::class, "destroyTeacher"])->name('deleteTeacher');
 
 
 
@@ -92,15 +97,45 @@ Route::put('/contacts/{contact}', [Contact2Controller::class, 'update'])->name('
 });
 
 
+// general guard dashboard ::::::::::::::::::::::::::::::::::::::::::::::::::::::::
+
+Route::group(['middleware' => ['auth', GeneralGuardMiddleware::class]], function () {
+
+    Route::delete('/dashboard/deleteMatiere/{id}', [directorController::class, "destroyMatiere"])->name('deleteMatiere');
+    Route::get('/dashboard/addMatiere', [directorController::class, "showCreateMatiere"])->name('addMatiere');
+    Route::post('/dashboard/addMatiere/', [directorController::class, "saveMatiere"])->name('saveMatiere');
+    Route::get('/dashboard/Matieres', [directorController::class, "displayMatieres"])->name('displayMatieres');
 
 
-// Route::group(['middleware' => ['auth', GeneralGuardMiddleware::class]], function () {
-//     // Route::get('/general-guard/dashboard', 'GeneralGuardController@dashboard')->name('general_guard.dashboard');
-// });
+
+    Route::get('/dashboard/departements', [directorController::class, "showDepartements"])->name('showDepartements');
+    Route::get('/dashboard/addDepartement', [directorController::class, "addDepartement"])->name('addDepartement');
+    Route::post('/dashboard/addDepartement', [directorController::class, "saveDepartement"])->name('saveDepartement');
+    Route::delete('/dashboard/deleteDepartement/{id}', [directorController::class, "destroyDepartement"])->name('deleteDepartement');
+
+
+
+    Route::get('/dashboard/professeurs', [directorController::class, "displayTeachers"])->name('displayTeachers');
+    Route::get('/dashboard/addTeacher', [directorController::class, "addTeacher"])->name('addTeacher');
+    Route::get('/dashboard/updateTeacher/{id}', [directorController::class, "updateTeacher"])->name('updateTeacher');
+    Route::patch('/dashboard/updateTeacher/{id}', [directorController::class, "saveUpdate"])->name('saveUpdateTeacher');
+    Route::post('/dashboard/addTeacher', [directorController::class, "saveTeacher"])->name('saveTeacher');
+    Route::delete('/dashboard/deleteTeacher/{id}', [directorController::class, "destroyTeacher"])->name('deleteTeacher');
+
+
+
+    Route::resource('news', NewsController::class)->names(['index' => 'siteDashboard']);
+    Route::resource('utilisations', UtilisationDuTempsController::class);
+});
+
+
+// teacher dashboard :::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
 
 Route::group(['middleware' => ['auth', TeacherMiddleware::class]], function () {
-    // Route::get('/teacher/dashboard', 'TeacherController@dashboard')->name('teacher.dashboard');
-    Route::get('/teacher/dashboard', [LoginController::class, "teacherDashboard"])->name('teacherDashboard');
+    Route::get('/teacher/dashboard', [teacherController::class, "teacherDashboard"])->name('teacherDashboard');
+    Route::match(['get', 'post'], '/teacher/absence/', [teacherController::class, "displayAbsence"])->name('displayAbsence');
+
+
 });
 
 
