@@ -17,7 +17,7 @@ class GeneralGuardMiddleware
      */
     public function handle(Request $request, Closure $next)
     {
-        if (Auth::user()->role === 'director' || Auth::user()->role === 'general_guard')
+        if ( Auth::user()->role === 'general_guard' || Auth::user()->role === 'director')
 
         {
             return $next($request);
