@@ -4,6 +4,8 @@ namespace App\Models;
 
 use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
+use App\Models\Teacher;
+
 
 class departement extends Model
 {
@@ -11,6 +13,11 @@ class departement extends Model
     protected $fillable = [
         'name',
         'matiere_id',
-        'departement_id'
+        'departement_id',
+        'teacher_id'
     ];
+    public function teachers()
+    {
+        return $this->belongsToMany(Teacher::class, 'teacher_departments');
+    }
 }
