@@ -1,5 +1,3 @@
-
-
 document.addEventListener('DOMContentLoaded', function () {
     var images = [
         '/images/269713058_427831879049165_7153829545722713717_n.jpeg',
@@ -9,7 +7,6 @@ document.addEventListener('DOMContentLoaded', function () {
 
     var currentIndex = 0;
     var homeSection = document.getElementById('acceuil');
-
 
     function changeBackgroundImage() {
         homeSection.style.backgroundImage = "url('" + images[currentIndex] + "')";
@@ -29,18 +26,34 @@ document.addEventListener('DOMContentLoaded', function () {
     }
 
     function startSlideshow() {
-        setInterval(nextBackgroundImage, 4000); // Change every 3 seconds
+        setInterval(nextBackgroundImage, 4000); // Change every 4 seconds
     }
 
     startSlideshow(); // Start slideshow on page load
 
     document.getElementById('prevBtn').addEventListener('click', prevBackgroundImage);
     document.getElementById('nextBtn').addEventListener('click', nextBackgroundImage);
+
+    const scrollUp = () => {
+        const scrollUp = document.getElementById('scroll-up');
+        if (window.scrollY >= 350) {
+            scrollUp.classList.add('show-scroll');
+        } else {
+            scrollUp.classList.remove('show-scroll');
+        }
+    };
+
+    window.addEventListener('scroll', scrollUp);
+
+    const msgeButton = document.querySelector(".togle");
+    const discussionBox = document.querySelector(".discussion");
+
+    msgeButton.addEventListener('click', () => {
+        discussionBox.classList.toggle('active');
+    });
 });
 
 
-
-//++++++++++++++++++++++++++++++++
 var slideIndex = 1;
 showSlides(slideIndex);
 
@@ -56,18 +69,22 @@ function showSlides(n) {
     var i;
     var slides = document.getElementsByClassName("mySlides");
     var dots = document.getElementsByClassName("dot");
-    if (n > slides.length) { slideIndex = 1; }
-    if (n < 1) { slideIndex = slides.length; }
+    if (n > slides.length) {
+        slideIndex = 1;
+    }
+    if (n < 1) {
+        slideIndex = slides.length;
+    }
     for (i = 0; i < slides.length; i++) {
         slides[i].style.display = "none";
     }
     for (i = 0; i < dots.length; i++) {
         dots[i].className = dots[i].className.replace(" active", "");
     }
-    slides[slideIndex - 1].style.display = "block";
-    dots[slideIndex - 1].className += " active";
+    if (slides[slideIndex - 1]) {
+        slides[slideIndex - 1].style.display = "block";
+    }
+    if (dots[slideIndex - 1]) {
+        dots[slideIndex - 1].className += " active";
+    }
 }
-
-
-
-
