@@ -10,9 +10,8 @@
             <p> Bienvenue à l'école ! Préparez-vous pour un nouveau chapitre passionnant
                 rempli d'opportunités de croissance et de succès. Ensemble, embarquons
                 dans cette aventure et faisons de cette année scolaire un moment inoubliable !</p>
-            <div>
-                <button class="but1">Lire Plus<i class="ri-arrow-right-up-fill"></i></i>
-                </button>
+            <div class="but22">
+                <button class="but1">Lire Plus<i class="ri-arrow-right-up-fill"></i></button>
                 <button class="but2">Nouvelles<i class="ri-news-line"></i></button>
             </div>
         </div>
@@ -63,80 +62,31 @@
         <p>Explorez les différentes voies éducatives offertes dans notre lycée pour trouver celle qui correspond le mieux à
             vos aspirations.</p>
         <div class="filiers">
-            <div class="filier-box">
-                <img src="images/pexels-jeswin-thomas-3781338.jpg" class="imaf">
-                <div class="details">
-                    <h6>Sciences Mathématiques <i class="ri-ruler-line"></i></h6>
-                    <p>Cette filière met l'accent sur l'étude des mathématiques et des sciences physiques, offrant aux
-                        étudiants une solide base théorique et pratique pour poursuivre des études supérieures en
-                        ingénierie, sciences, ou mathématiques appliquées. </p>
 
-                </div>
+
+            <div class="filier-box">
+                <img src="images/Sciences Mathématiques (5).png" class="imaf">
+            </div>
+
+            <div class="filier-box">
+                <img src="images/Sciences Mathématiques (1).png" class="imaf">
+            </div>
+
+            <div class="filier-box">
+                <img src="images/Sciences Mathématiques (2).png" class="imaf">
+            </div>
+
+            <div class="filier-box">
+                <img src="images/Sciences Mathématiques (3).png" class="imaf">
             </div>
 
 
             <div class="filier-box">
-                <img src="images/pexels-lucas-pezeta-5277388.jpg" class="imaf">
-                <div class="details">
-                    <h6>Sciences de la Vie et la Terre <i class="ri-plant-line"></i></h6>
-                    <p>Cette filière explore les sciences naturelles, y compris la biologie, la géologie et l'écologie,
-                        offrant aux étudiants une compréhension approfondie du monde vivant et de l'environnement, avec
-                        des perspectives sur la santé, l'agriculture et la conservation.</p>
-
-                </div>
+                <img src="images/Sciences Mathématiques (4).png" class="imaf">
             </div>
 
             <div class="filier-box">
-                <img src="images/pexels-chokniti-khongchum-2280571.jpg" class="imaf">
-                <div class="details">
-                    <h6>Sciences Physiques Chimiques <i class="ri-ink-bottle-line"></i> </h6>
-                    <p> pour Cette filière se concentre sur l'étude approfondie de la physique et de la chimie développant
-                        les
-                        compétences essentielles en analyse et expérimentation pour l'ingénierie,la recherche scientifique
-                        et
-                        la Physique-chimie essentiel pour les étudiants. </p>
-
-                </div>
-            </div>
-
-            <div class="filier-box">
-                <img src="images/pexels-pixabay-53621.jpg" class="imaf">
-                <div class="details">
-                    <h6>Sciences Économiques <i class="ri-calculator-line"></i></h6>
-                    <p>Exploration approfondie des théories économiques et des concepts de gestion, préparant les
-                        étudiants à des carrières variées dans le domaine financier, commercial Acquisition de compétences
-                        analytiques essentielles pour comprendre
-                        les dynamiques économiques contemporaines </p>
-
-                </div>
-            </div>
-
-
-            <div class="filier-box">
-                <img src="images/pexels-jean-marc-bonnel-19081217.jpg" class="imaf">
-                <div class="details">
-                    <h6>Sciences Humaines <i class="ri-globe-line"></i> </h6>
-                    <p>Étude approfondie de disciplines telles que l'histoire,
-                        la géographie et la sociologie, offrant une perspective critique sur la société et la culture .
-                        Cette exploration permet aux étudiants de comprendre les dynamiques sociales,
-                        historiques et géographiques du pays.</p>
-
-                </div>
-            </div>
-
-
-
-            <div class="filier-box">
-                <img src="images/section-litteraire-lycee.jpeg" class="imaf">
-                <div class="details">
-                    <h6>Baccalauréat Lettres <i class="ri-git-repository-line"></i></h6>
-                    <p>Programme mettant l'accent sur l'étude approfondie de la littérature,
-                        de la philosophie, de l'histoire et des langues vivantes, préparant les étudiants à des études
-                        supérieures en lettres, en sciences humaines et dans d'autres domaines connexes Enseignement riche,
-                        orienté littérature,
-                        histoire </p>
-
-                </div>
+                <img src="images/Sciences Mathématiques.png" class="imaf">
             </div>
         </div>
     </section>
@@ -155,7 +105,6 @@
         <div class="slideshow-container">
             <div class="mySlides">
                 <i class="ri-user-fill user"></i>
-
                 <p class="card__description">
                     "J'ai trouvé au lycée un environnement propice à
                     l'épanouissement personnel et académique. Les opportunités
@@ -168,7 +117,6 @@
 
             <div class="mySlides">
                 <i class="ri-user-fill user"></i>
-
                 <p class="card__description">
                     "Le lycée m'a donné les outils nécessaires pour atteindre mes
                     objectifs académiques et professionnels. Les enseignants dévoués et les
@@ -180,7 +128,6 @@
 
             <div class="mySlides">
                 <i class="ri-user-fill user"></i>
-
                 <p class="card__description">
                     "Les diverses activités parascolaires proposées par
                     le lycée ont enrichi mon expérience éducative. J'ai
@@ -193,7 +140,6 @@
 
             <div class="mySlides">
                 <i class="ri-user-fill user"></i>
-
                 <p class="card__description">
                     "Mon parcours au lycée a été marqué par des rencontres
                     inspirantes et des défis stimulants. J'ai été soutenu par
@@ -208,6 +154,9 @@
 
 
 
+
+
+
             <div style="text-align:center">
                 <span class="dot" onclick="currentSlide(1)"></span>
                 <span class="dot" onclick="currentSlide(2)"></span>
@@ -216,10 +165,16 @@
 
             </div>
 
+
         </div>
 
 
+
     </section>
+
+
+
+
 
 
 
@@ -227,8 +182,61 @@
 
     <!-- contact -->
     <section id="contact">
+        <h1>Envoie-nous un message</h1>
+        <p>Besoin d'aide ? Laissez-nous un message ! Nous sommes là pour vous aider avec toutes vos questions et
+            préoccupations.</p>
+
+
+    <section id="contact">
     <h1>Envoie-nous un message</h1>
     <p>Besoin d'aide ? Laissez-nous un message ! Nous sommes là pour vous aider avec toutes vos questions et préoccupations.</p>
+
+        <div class="container">
+            <div class="content">
+                <div class="left-side">
+                    <div class="address details">
+                        <i class="ri-map-pin-line"></i>
+                        <div class="topic">Adresse</div>
+                        <div class="text-one">Driouch</div>
+                        <div class="text-two">HAY AL AMAL 06</div>
+                    </div>
+                    <div class="phone details">
+                        <i class="ri-phone-line"></i>
+                        <div class="topic">Téléphone</div>
+                        <div class="text-one">+212 70652357</div>
+                        <div class="text-two">+212 62OO21O1</div>
+                    </div>
+                    <div class="email details">
+                        <i class="ri-mail-line"></i>
+                        <div class="topic">Email</div>
+                        <div class="text-one">moulay_ismail@gmail.com</div>
+                        <div class="text-two">info_ms@gmail.com</div>
+                    </div>
+                </div>
+                <div class="right-side">
+                    <div class="topic-text">Contact</div>
+                    <p>Si vous avez un travail de ma part ou tout type de requêtes liées à mon tutoriel, vous pouvez
+                        m'envoyer un message à partir d'ici. C'est avec plaisir que je vous aide.</p>
+                    <form action="#">
+                        <div class="input-box">
+                            <input type="text" placeholder="Entrez votre nom">
+                        </div>
+                        <div class="input-box">
+                            <input type="text" placeholder="Entrer votre Email">
+                        </div>
+                        <div class="input-box message-box">
+                            <input type="texteria" placeholder=" Entrez votre message">
+                        </div>
+                        <div class="button">
+                            <input type="button" value="Envoyer">
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </div>
+
+
+   </section>
 
     <div class="container">
         <div class="content">
@@ -284,6 +292,65 @@
 
      <!--  end contact -->
 
+
+    <!-- Button to scroll back to top -->
+    <a href="#" class="scrollup" id="scroll-up"> <i class="ri-arrow-up-s-line"></i> </a>
+
+
+    <!-- togl -->
+
+    <div class="discussion">
+        <div class="part1">
+            <h2>Ajouter Vos Témoignages</h2>
+         </div>
+
+
+        <div class="part2">
+
+                 <p>
+                    "Exprimez-vous ! Partagez vos expériences uniques avec nous.
+                    Votre témoignage pourrait inspirer d'autres."
+                </p>
+        </div>
+
+        <dir class="part3">
+            <form class="formul-togl">
+
+          <div class="inps">
+            <table>
+            <tr>
+                  <td> <label for="a" class="i1"><i class="ri-id-card-line"></i></label></td>
+                   <td> <input id="a" type="text" name="nom" class="n1" required placeholder="Entrez votre nom"></td>
+
+            </tr>
+            <tr>
+                <td> <label for="b" class="i1"><i class="ri-calendar-2-line"></i></label></td>
+                 <td> <input id="b" type="text" name="time" class="n1" required placeholder="Année scolaire"></td>
+
+            </tr>
+
+            </table>
+          </div>
+
+
+
+           <div class="msg">
+           <input class="n2" type="text" placeholder="Entrez votre Témoignages ici...">
+           <button type="submit" name="ajouter" class="n3"><i class="ri-send-plane-fill"></i></button>
+
+           </div>
+
+       </form>
+
+
+        </dir>
+
+    </div>
+
+
+    <div class="articlex1">
+        <button class="togle"><i class="ri-message-3-line"></i></button>
+    </div>
 
 
 
